requirements.txt
fastapi
pandas
torch
transformers
Pillow
uvicorn
# Set up the environment
git+https://github.com/xinntao/Real-ESRGAN.git
git+https://github.com/XPixelGroup/BasicSR.git